refactor(session): extract detectRepetition to shared util
Move detectRepetition function and constants from processor.ts to
a dedicated repetition.ts module. Test now imports production code
instead of reimplementing the algorithm.

Closes #69 (partial)
