+++ packages/opencode/src/session/processor.ts
@@ -19,12 +19,10 @@ import { SessionStatus } from "./status"
 import { SessionSummary } from "./summary"
 import type { Provider } from "@/provider/provider"
 import { Question } from "@/question"
+import { detectRepetition, REPETITION_THRESHOLD } from "./repetition"
 
 export namespace SessionProcessor {
   const DOOM_LOOP_THRESHOLD = 3
-  // 50+ consecutive repeats of a 4-200 char pattern in the last 8KB indicates a model generation loop
-  const REPETITION_THRESHOLD = 50
-  const REPETITION_WINDOW = 8000
   const log = Log.create({ service: "session.processor" })
 
   class RepetitionError extends Error {
@@ -32,24 +30,6 @@ export namespace SessionProcessor {
       super(`Repetition loop detected: "${pattern.substring(0, 80)}..." repeated ${REPETITION_THRESHOLD}+ times. Aborting to prevent context exhaustion.`)
       this.name = "RepetitionError"
     }
-  }
-
-  function detectRepetition(text: string): boolean {
-    if (text.length < REPETITION_WINDOW) return false
-    const tail = text.slice(-REPETITION_WINDOW)
-    for (let len = 4; len <= 200; len++) {
-      const pattern = tail.slice(-len)
-      let count = 0
-      let pos = tail.length - len
-      while (pos >= 0) {
-        if (tail.slice(pos, pos + len) === pattern) {
-          count++
-          pos -= len
-        } else break
-      }
-      if (count >= REPETITION_THRESHOLD) return true
-    }
-    return false
   }
 
   export type Result = "compact" | "stop" | "continue"
